explicitly add '*' as the allowed-origins, and do not use existing CORS handling, which does not work well with web-caching (and we have had no success in getting www-prod to alter the rules for this either).
