+++ rest-common/src/test/java/uk/ac/ebi/quickgo/rest/controller/AllowAllOriginsResponseFilterIT.java
@@ -1,0 +1,81 @@
+package uk.ac.ebi.quickgo.rest.controller;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.http.HttpHeaders;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static org.springframework.http.HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static uk.ac.ebi.quickgo.rest.controller.AllowAllOriginsResponseFilter.ALLOW_ALL_ORIGINS;
+import static uk.ac.ebi.quickgo.rest.controller.FakeRESTApp.RESOURCE_1_URL;
+
+/**
+ * Check that a REST app that picks up an {@link AllowAllOriginsResponseFilter} will show explicitly
+ * the all origins header value, '*' -- and not the actual origin.
+ * Created 13/03/18
+ * @author Edd
+ */
+@ActiveProfiles("allow-origins-integration-test")
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringApplicationConfiguration(classes = {FakeRESTApp.class})
+@WebAppConfiguration
+public class AllowAllOriginsResponseFilterIT {
+    @Autowired
+    private WebApplicationContext webApplicationContext;
+    private MockMvc mockMvc;
+    @Autowired
+    private AllowAllOriginsResponseFilter responseFilter;
+
+    @Before
+    public void setUp() {
+        mockMvc = MockMvcBuilders
+                .webAppContextSetup(webApplicationContext)
+                .addFilter(responseFilter)
+                .build();
+    }
+
+    @Test
+    public void requestWithoutAnOriginHasResponseWithAllOriginsHeader() throws Exception {
+        MvcResult result = mockMvc.perform(
+                get(RESOURCE_1_URL))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andReturn();
+
+        assertThat(result.getResponse().getHeader(ACCESS_CONTROL_ALLOW_ORIGIN), is(ALLOW_ALL_ORIGINS));
+    }
+
+    @Test
+    public void requestFromAnOriginHasResponseWithAllOriginsHeader() throws Exception {
+        String origin = "http://www.ebi.ac.uk";
+
+        MvcResult result = mockMvc.perform(
+                get(RESOURCE_1_URL)
+                        .headers(originHeader(origin)))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andReturn();
+
+        assertThat(result.getResponse().getHeader(ACCESS_CONTROL_ALLOW_ORIGIN), is(ALLOW_ALL_ORIGINS));
+    }
+
+    private HttpHeaders originHeader(String origin) {
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setOrigin(origin);
+        return httpHeaders;
+    }
+}
